refactor: CustomMui 및 icon의 하위 폴더 제거
