+++ src/pages/Search/components/CurrentRegionButton/index.tsx
@@ -1,5 +1,3 @@
-import CustomButton from '@/components/CustomMui/CustomButton';
-import LocationIcon from '@/components/icon/Location';
 import { S } from './style';
 import useAppSelector from '@/hooks/useAppSelector';
 import useAppDispatch from '@/hooks/useAppDispatch';
@@ -7,6 +5,8 @@ import { currentRegionActions } from '@/redux/slice/currentRegionSlice';
 import { useNavigate } from 'react-router-dom';
 import { MY_REGION } from '@/constants/localStorage/key';
 import useGeolocation from '@/hooks/useGeolocation';
+import CustomButton from '@/components/CustomMui/CustomButton';
+import LocationIcon from '@/components/icon/Location';
 
 const CurrentRegionButton = () => {
   const currentRegion = useAppSelector((state) => state.currentRegion.value);
