+++ src/pages/Search/components/SearchHeader/index.tsx
@@ -1,10 +1,10 @@
-import CustomPaper from '@/components/CustomMui/CustomPaper';
 import { C, S } from './style';
-import CustomToolbar from '@/components/CustomMui/CustomToolBar';
 import { Link } from 'react-router-dom';
 import ArrowIcon from '@/assets/svg/arrow.svg?react';
-import CustomTextField from '@/components/CustomMui/CustomTextField';
 import CancelIcon from '@/assets/svg/cancle.svg?react';
+import CustomPaper from '@/components/CustomMui/CustomPaper';
+import CustomToolbar from '@/components/CustomMui/CustomToolbar';
+import CustomTextField from '@/components/CustomMui/CustomTextField';
 
 type SearchHeaderProps = {
   keyword: string;
